requirements.txt
python-telegram-bot==20.7
yt-dlp==2024.1.7
python-dotenv==1.0.0

